updated simulator and boards

+++ trunk/f10_4444_mosquito_v1/src/mosquito/sim/Collector.java
@@ -27,5 +27,5 @@ public class Collector extends GameObject {
 	{
 		return boundingBox.contains(p);
 	}
-	public static final int RADIUS = 1;
+	public static final double RADIUS = 1;
 }
